Updating client to handle callbacks in a more creative way.
Now, when subscription requests are made, a random callback URI
is generated and served by a dynamically-allocated http server.

Still need to work out how to close the callback server.  Perhaps
we should have a single long-running listenAndServe, but dynamic handler
patterns?  Something to think about.  I'm not sure if you can add and
deduct patterns on-the-fly.